create entities

+++ Entities/Concrete/Basket.cs
@@ -5,16 +5,14 @@ using System.Text;
 
 namespace Entities.Concrete
 {
-    public class Basket :IEntity
+    public class Basket : IEntity
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int ProductId { get; set; }
         public int UserId { get; set; }
         public int Count { get; set; }
         public DateTime CreateDate { get; set; }
         public bool Active { get; set; }
-
-
     }
 }
